Separate route logic into controllers
